requirements.txt
caldav==1.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
icalendar==6.1.3
idna==3.10
lxml==5.3.2
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
recurring-ical-events==3.7.0
requests==2.32.3
RPi.GPIO==0.7.1
six==1.17.0
smbus2==0.5.0
tzdata==2025.2
urllib3==2.4.0
vobject==0.9.9
x_wr_timezone==2.0.1
